+++ rust/crd/src/lib.rs
@@ -11,28 +11,30 @@ use crate::authentication::DruidAuthentication;
 use crate::tls::DruidTls;
 
 use authorization::DruidAuthorization;
+use resource::RoleResource;
 use serde::{Deserialize, Serialize};
 use snafu::{OptionExt, ResultExt, Snafu};
 use stackable_operator::{
     client::Client,
     commons::{
         product_image_selection::ProductImage,
-        resources::{NoRuntimeLimits, ResourcesFragment},
+        resources::{NoRuntimeLimits, Resources},
         s3::{InlinedS3BucketSpec, S3BucketDef, S3ConnectionDef, S3ConnectionSpec},
         tls::{CaCert, Tls, TlsServerVerification, TlsVerification},
+    },
+    config::{
+        fragment::{self, Fragment, FromFragment, ValidationError},
+        merge::Merge,
     },
     k8s_openapi::apimachinery::pkg::apis::meta::v1::LabelSelector,
     kube::{CustomResource, ResourceExt},
     labels::ObjectLabels,
     product_config::types::PropertyNameKind,
     product_config_utils::{ConfigError, Configuration},
-    role_utils::{Role, RoleGroupRef},
+    role_utils::{CommonConfiguration, Role, RoleGroup},
     schemars::{self, JsonSchema},
 };
-use std::{
-    collections::{BTreeMap, HashMap},
-    str::FromStr,
-};
+use std::collections::{BTreeMap, HashMap};
 use strum::{Display, EnumDiscriminants, EnumIter, EnumString, IntoStaticStr};
 use tls::default_druid_tls;
 
@@ -123,10 +125,12 @@ pub enum Error {
     },
     #[snafu(display("2 differing s3 connections were given, this is unsupported by Druid"))]
     IncompatibleS3Connections,
-    #[snafu(display("Unknown Druid role found {role}. Should be one of {roles:?}"))]
-    UnknownDruidRole { role: String, roles: Vec<String> },
+    #[snafu(display("the role group {rolegroup_name} is not defined"))]
+    CannotRetrieveRoleGroup { rolegroup_name: String },
     #[snafu(display("missing namespace for resource {name}"))]
     MissingNamespace { name: String },
+    #[snafu(display("fragment validation failure"))]
+    FragmentValidationFailure { source: ValidationError },
 }
 
 #[derive(Clone, CustomResource, Debug, Deserialize, JsonSchema, Serialize)]
@@ -152,15 +156,15 @@ pub struct DruidClusterSpec {
     /// The Druid image to use
     pub image: ProductImage,
     /// Configuration of the broker role
-    pub brokers: Role<BrokerConfig>,
+    pub brokers: Role<BrokerConfigFragment>,
     /// Configuration of the coordinator role
-    pub coordinators: Role<CoordinatorConfig>,
+    pub coordinators: Role<CoordinatorConfigFragment>,
     /// Configuration of the historical role
-    pub historicals: Role<HistoricalConfig>,
+    pub historicals: Role<HistoricalConfigFragment>,
     /// Configuration of the middle managed role
-    pub middle_managers: Role<MiddleManagerConfig>,
+    pub middle_managers: Role<MiddleManagerConfigFragment>,
     /// Configuration of the router role
-    pub routers: Role<RouterConfig>,
+    pub routers: Role<RouterConfigFragment>,
     /// Common cluster wide configuration that can not differ or be overridden on a role or role group level
     pub cluster_config: DruidClusterConfig,
 }
@@ -188,6 +192,100 @@ pub struct DruidClusterConfig {
     pub tls: Option<DruidTls>,
     /// ZooKeeper discovery ConfigMap
     pub zookeeper_config_map_name: String,
+}
+
+/// Common configuration for all role groups
+pub struct CommonRoleGroupConfig {
+    pub resources: RoleResource,
+    pub replicas: Option<u16>,
+    pub selector: Option<LabelSelector>,
+}
+
+/// Container for the merged and validated role group configurations
+///
+/// This structure contains for every role a map from the role group names to their configurations.
+/// The role group configurations are merged with the role and default configurations. The product
+/// configuration is not applied.
+pub struct MergedConfig {
+    /// Merged configuration of the broker role
+    pub brokers: HashMap<String, RoleGroup<BrokerConfig>>,
+    /// Merged configuration of the coordinator role
+    pub coordinators: HashMap<String, RoleGroup<CoordinatorConfig>>,
+    /// Merged configuration of the historical role
+    pub historicals: HashMap<String, RoleGroup<HistoricalConfig>>,
+    /// Merged configuration of the middle manager role
+    pub middle_managers: HashMap<String, RoleGroup<MiddleManagerConfig>>,
+    /// Merged configuration of the router role
+    pub routers: HashMap<String, RoleGroup<RouterConfig>>,
+}
+
+impl MergedConfig {
+    /// Returns the common configuration for the given role and rolegroup name
+    pub fn common_config(
+        &self,
+        role: DruidRole,
+        rolegroup_name: &str,
+    ) -> Result<CommonRoleGroupConfig, Error> {
+        match role {
+            DruidRole::Broker => {
+                let rolegroup = self
+                    .brokers
+                    .get(rolegroup_name)
+                    .context(CannotRetrieveRoleGroupSnafu { rolegroup_name })?;
+                Ok(CommonRoleGroupConfig {
+                    resources: RoleResource::Druid(rolegroup.config.config.resources.to_owned()),
+                    replicas: rolegroup.replicas,
+                    selector: rolegroup.selector.to_owned(),
+                })
+            }
+            DruidRole::Coordinator => {
+                let rolegroup = self
+                    .coordinators
+                    .get(rolegroup_name)
+                    .context(CannotRetrieveRoleGroupSnafu { rolegroup_name })?;
+                Ok(CommonRoleGroupConfig {
+                    resources: RoleResource::Druid(rolegroup.config.config.resources.to_owned()),
+                    replicas: rolegroup.replicas,
+                    selector: rolegroup.selector.to_owned(),
+                })
+            }
+            DruidRole::Historical => {
+                let rolegroup = self
+                    .historicals
+                    .get(rolegroup_name)
+                    .context(CannotRetrieveRoleGroupSnafu { rolegroup_name })?;
+                Ok(CommonRoleGroupConfig {
+                    resources: RoleResource::Historical(
+                        rolegroup.config.config.resources.to_owned(),
+                    ),
+                    replicas: rolegroup.replicas,
+                    selector: rolegroup.selector.to_owned(),
+                })
+            }
+            DruidRole::MiddleManager => {
+                let rolegroup = self
+                    .middle_managers
+                    .get(rolegroup_name)
+                    .context(CannotRetrieveRoleGroupSnafu { rolegroup_name })?;
+                Ok(CommonRoleGroupConfig {
+                    resources: RoleResource::Druid(rolegroup.config.config.resources.to_owned()),
+                    replicas: rolegroup.replicas,
+                    selector: rolegroup.selector.to_owned(),
+                })
+            }
+            DruidRole::Router => {
+                let rolegroup = self
+                    .routers
+                    .get(rolegroup_name)
+                    .context(CannotRetrieveRoleGroupSnafu { rolegroup_name })?;
+                Ok(CommonRoleGroupConfig {
+                    resources: RoleResource::Druid(rolegroup.config.config.resources.to_owned()),
+                    replicas: rolegroup.replicas,
+                    selector: rolegroup.selector.to_owned(),
+                })
+            }
+        }
+    }
 }
 
 #[derive(
@@ -371,86 +469,6 @@ impl DruidCluster {
         Ok(result)
     }
 
-    /// Takes a rolegoup_ref (with role and role group name) and returns the selector defined for
-    /// that role group.
-    pub fn node_selector(
-        &self,
-        rolegroup_ref: &RoleGroupRef<DruidCluster>,
-    ) -> Option<LabelSelector> {
-        match DruidRole::from_str(rolegroup_ref.role.as_str()).unwrap() {
-            DruidRole::Broker => self
-                .spec
-                .brokers
-                .role_groups
-                .get(&rolegroup_ref.role_group)
-                .and_then(|rg| rg.selector.clone()),
-            DruidRole::MiddleManager => self
-                .spec
-                .middle_managers
-                .role_groups
-                .get(&rolegroup_ref.role_group)
-                .and_then(|rg| rg.selector.clone()),
-            DruidRole::Coordinator => self
-                .spec
-                .coordinators
-                .role_groups
-                .get(&rolegroup_ref.role_group)
-                .and_then(|rg| rg.selector.clone()),
-            DruidRole::Historical => self
-                .spec
-                .historicals
-                .role_groups
-                .get(&rolegroup_ref.role_group)
-                .and_then(|rg| rg.selector.clone()),
-            DruidRole::Router => self
-                .spec
-                .routers
-                .role_groups
-                .get(&rolegroup_ref.role_group)
-                .and_then(|rg| rg.selector.clone()),
-        }
-    }
-
-    pub fn replicas(&self, rolegroup_ref: &RoleGroupRef<DruidCluster>) -> Option<i32> {
-        match DruidRole::from_str(rolegroup_ref.role.as_str()).unwrap() {
-            DruidRole::Broker => self
-                .spec
-                .brokers
-                .role_groups
-                .get(&rolegroup_ref.role_group)
-                .and_then(|rg| rg.replicas)
-                .map(i32::from),
-            DruidRole::MiddleManager => self
-                .spec
-                .middle_managers
-                .role_groups
-                .get(&rolegroup_ref.role_group)
-                .and_then(|rg| rg.replicas)
-                .map(i32::from),
-            DruidRole::Coordinator => self
-                .spec
-                .coordinators
-                .role_groups
-                .get(&rolegroup_ref.role_group)
-                .and_then(|rg| rg.replicas)
-                .map(i32::from),
-            DruidRole::Historical => self
-                .spec
-                .historicals
-                .role_groups
-                .get(&rolegroup_ref.role_group)
-                .and_then(|rg| rg.replicas)
-                .map(i32::from),
-            DruidRole::Router => self
-                .spec
-                .routers
-                .role_groups
-                .get(&rolegroup_ref.role_group)
-                .and_then(|rg| rg.replicas)
-                .map(i32::from),
-        }
-    }
-
     pub fn build_role_properties(
         &self,
     ) -> HashMap<
@@ -591,6 +609,98 @@ impl DruidCluster {
         let s3_storage = self.spec.cluster_config.deep_storage.is_s3();
         s3_ingestion || s3_storage
     }
+
+    /// Returns the merged and validated configuration for all roles
+    pub fn merged_config(&self) -> Result<MergedConfig, Error> {
+        Ok(MergedConfig {
+            brokers: DruidCluster::merged_role(
+                &self.spec.brokers,
+                &BrokerConfig::default_config(),
+            )?,
+            coordinators: DruidCluster::merged_role(
+                &self.spec.coordinators,
+                &CoordinatorConfig::default_config(),
+            )?,
+            historicals: DruidCluster::merged_role(
+                &self.spec.historicals,
+                &HistoricalConfig::default_config(),
+            )?,
+            middle_managers: DruidCluster::merged_role(
+                &self.spec.middle_managers,
+                &MiddleManagerConfig::default_config(),
+            )?,
+            routers: DruidCluster::merged_role(
+                &self.spec.routers,
+                &RouterConfig::default_config(),
+            )?,
+        })
+    }
+
+    /// Merges and validates the role groups of the given role with the given default configuration
+    fn merged_role<T>(
+        role: &Role<T::Fragment>,
+        default_config: &T::Fragment,
+    ) -> Result<HashMap<String, RoleGroup<T>>, Error>
+    where
+        T: FromFragment,
+        T::Fragment: Clone + Merge,
+    {
+        let mut merged_role_config = HashMap::new();
+
+        for (rolegroup_name, rolegroup) in &role.role_groups {
+            let merged_rolegroup_config =
+                DruidCluster::merged_rolegroup(rolegroup, &role.config.config, default_config)?;
+            merged_role_config.insert(rolegroup_name.to_owned(), merged_rolegroup_config);
+        }
+
+        Ok(merged_role_config)
+    }
+
+    /// Merges and validates the given role group with the given role and default configurations
+    fn merged_rolegroup<T>(
+        rolegroup: &RoleGroup<T::Fragment>,
+        role_config: &T::Fragment,
+        default_config: &T::Fragment,
+    ) -> Result<RoleGroup<T>, Error>
+    where
+        T: FromFragment,
+        T::Fragment: Clone + Merge,
+    {
+        let merged_config = DruidCluster::merged_rolegroup_config(
+            &rolegroup.config.config,
+            role_config,
+            default_config,
+        )?;
+        Ok(RoleGroup {
+            config: CommonConfiguration {
+                config: merged_config,
+                config_overrides: rolegroup.config.config_overrides.to_owned(),
+                env_overrides: rolegroup.config.env_overrides.to_owned(),
+                cli_overrides: rolegroup.config.cli_overrides.to_owned(),
+            },
+            replicas: rolegroup.replicas,
+            selector: rolegroup.selector.to_owned(),
+        })
+    }
+
+    /// Merges and validates the given role group, role, and default configurations
+    pub fn merged_rolegroup_config<T>(
+        rolegroup_config: &T::Fragment,
+        role_config: &T::Fragment,
+        default_config: &T::Fragment,
+    ) -> Result<T, Error>
+    where
+        T: FromFragment,
+        T::Fragment: Clone + Merge,
+    {
+        let mut role_config = role_config.to_owned();
+        let mut rolegroup_config = rolegroup_config.to_owned();
+
+        role_config.merge(default_config);
+        rolegroup_config.merge(&role_config);
+
+        fragment::validate(rolegroup_config).context(FragmentValidationFailureSnafu)
+    }
 }
 
 #[derive(Clone, Debug, Default, Deserialize, JsonSchema, Serialize)]
@@ -664,37 +774,142 @@ pub struct IngestionSpec {
     pub s3connection: Option<S3ConnectionDef>,
 }
 
-#[derive(Clone, Debug, Default, Deserialize, JsonSchema, PartialEq, Serialize)]
-#[serde(rename_all = "camelCase")]
+#[derive(Clone, Debug, Default, Fragment, JsonSchema, PartialEq)]
+#[fragment_attrs(
+    derive(
+        Clone,
+        Debug,
+        Default,
+        Deserialize,
+        Merge,
+        JsonSchema,
+        PartialEq,
+        Serialize
+    ),
+    serde(rename_all = "camelCase")
+)]
 pub struct BrokerConfig {
-    resources: Option<ResourcesFragment<storage::DruidStorage, NoRuntimeLimits>>,
+    #[fragment_attrs(serde(default))]
+    resources: Resources<storage::DruidStorage, NoRuntimeLimits>,
 }
 
-#[derive(Clone, Debug, Default, Deserialize, JsonSchema, PartialEq, Serialize)]
-#[serde(rename_all = "camelCase")]
+impl BrokerConfig {
+    fn default_config() -> BrokerConfigFragment {
+        BrokerConfigFragment {
+            resources: resource::DEFAULT_RESOURCES.to_owned(),
+        }
+    }
+}
+
+#[derive(Clone, Debug, Default, Fragment, JsonSchema, PartialEq)]
+#[fragment_attrs(
+    derive(
+        Clone,
+        Debug,
+        Default,
+        Deserialize,
+        Merge,
+        JsonSchema,
+        PartialEq,
+        Serialize
+    ),
+    serde(rename_all = "camelCase")
+)]
 pub struct CoordinatorConfig {
-    resources: Option<ResourcesFragment<storage::DruidStorage, NoRuntimeLimits>>,
+    #[fragment_attrs(serde(default))]
+    resources: Resources<storage::DruidStorage, NoRuntimeLimits>,
 }
 
-#[derive(Clone, Debug, Default, Deserialize, JsonSchema, PartialEq, Serialize)]
-#[serde(rename_all = "camelCase")]
+impl CoordinatorConfig {
+    fn default_config() -> CoordinatorConfigFragment {
+        CoordinatorConfigFragment {
+            resources: resource::DEFAULT_RESOURCES.to_owned(),
+        }
+    }
+}
+
+#[derive(Clone, Debug, Default, Fragment, JsonSchema, PartialEq)]
+#[fragment_attrs(
+    derive(
+        Clone,
+        Debug,
+        Default,
+        Deserialize,
+        Merge,
+        JsonSchema,
+        PartialEq,
+        Serialize
+    ),
+    serde(rename_all = "camelCase")
+)]
 pub struct MiddleManagerConfig {
-    resources: Option<ResourcesFragment<storage::DruidStorage, NoRuntimeLimits>>,
+    #[fragment_attrs(serde(default))]
+    resources: Resources<storage::DruidStorage, NoRuntimeLimits>,
 }
 
-#[derive(Clone, Debug, Default, Deserialize, JsonSchema, PartialEq, Serialize)]
-#[serde(rename_all = "camelCase")]
+impl MiddleManagerConfig {
+    fn default_config() -> MiddleManagerConfigFragment {
+        MiddleManagerConfigFragment {
+            resources: resource::DEFAULT_RESOURCES.to_owned(),
+        }
+    }
+}
+
+#[derive(Clone, Debug, Default, Fragment, JsonSchema, PartialEq)]
+#[fragment_attrs(
+    derive(
+        Clone,
+        Debug,
+        Default,
+        Deserialize,
+        Merge,
+        JsonSchema,
+        PartialEq,
+        Serialize
+    ),
+    serde(rename_all = "camelCase")
+)]
 pub struct RouterConfig {
-    resources: Option<ResourcesFragment<storage::DruidStorage, NoRuntimeLimits>>,
+    #[fragment_attrs(serde(default))]
+    resources: Resources<storage::DruidStorage, NoRuntimeLimits>,
 }
 
-#[derive(Clone, Debug, Default, Deserialize, JsonSchema, PartialEq, Serialize)]
-#[serde(rename_all = "camelCase")]
+impl RouterConfig {
+    fn default_config() -> RouterConfigFragment {
+        RouterConfigFragment {
+            resources: resource::DEFAULT_RESOURCES.to_owned(),
+        }
+    }
+}
+
+#[derive(Clone, Debug, Default, Fragment, JsonSchema, PartialEq)]
+#[fragment_attrs(
+    derive(
+        Clone,
+        Debug,
+        Default,
+        Deserialize,
+        Merge,
+        JsonSchema,
+        PartialEq,
+        Serialize
+    ),
+    serde(rename_all = "camelCase")
+)]
 pub struct HistoricalConfig {
-    resources: Option<ResourcesFragment<storage::HistoricalStorage, NoRuntimeLimits>>,
+    #[fragment_attrs(serde(default))]
+    resources: Resources<storage::HistoricalStorage, NoRuntimeLimits>,
 }
 
-impl Configuration for BrokerConfig {
+impl HistoricalConfig {
+    fn default_config() -> HistoricalConfigFragment {
+        HistoricalConfigFragment {
+            resources: resource::HISTORICAL_RESOURCES.to_owned(),
+        }
+    }
+}
+
+impl Configuration for BrokerConfigFragment {
     type Configurable = DruidCluster;
 
     fn compute_env(
@@ -724,7 +939,7 @@ impl Configuration for BrokerConfig {
     }
 }
 
-impl Configuration for HistoricalConfig {
+impl Configuration for HistoricalConfigFragment {
     type Configurable = DruidCluster;
 
     fn compute_env(
@@ -754,7 +969,7 @@ impl Configuration for HistoricalConfig {
     }
 }
 
-impl Configuration for RouterConfig {
+impl Configuration for RouterConfigFragment {
     type Configurable = DruidCluster;
 
     fn compute_env(
@@ -784,7 +999,7 @@ impl Configuration for RouterConfig {
     }
 }
 
-impl Configuration for MiddleManagerConfig {
+impl Configuration for MiddleManagerConfigFragment {
     type Configurable = DruidCluster;
 
     fn compute_env(
@@ -823,7 +1038,7 @@ impl Configuration for MiddleManagerConfig {
     }
 }
 
-impl Configuration for CoordinatorConfig {
+impl Configuration for CoordinatorConfigFragment {
     type Configurable = DruidCluster;
 
     fn compute_env(